requirements.txt
numpy==1.24.4
pandas==2.2.3
Pillow==11.2.1
PyYAML==6.0.2
PyYAML==6.0.2
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.0
